Moved tag-header to components' folder. Removed SchoolHeader.
