+++ src/components/tag-header.js
@@ -18,12 +18,12 @@
 import React, { PropTypes } from 'react';
 import { Link, IndexLink } from 'react-router';
 
-import Panel from '../panel';
-import Time from '../time';
-import FollowTagButton from '../follow-tag-button';
-import Tag from '../tag';
-import LikeTagButton from '../like-tag-button';
-import { TAG_SCHOOL, TAG_LOCATION, TAG_HASHTAG } from '../../consts/tags';
+import Panel from './panel';
+import Time from './time';
+import FollowTagButton from './follow-tag-button';
+import Tag from './tag';
+import LikeTagButton from './like-tag-button';
+import { TAG_SCHOOL, TAG_LOCATION, TAG_HASHTAG } from '../consts/tags';
 
 function getLikeTriggers(triggers, type) {
   switch (type) {
